--- member/src/main/java/com/mossy/boundedContext/exception/SuccessCode.java
@@ -1,7 +1,0 @@
-package com.mossy.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public enum SuccessCode {
-}
