--- member/src/main/java/com/mossy/boundedContext/global/security/RestAccessDeniedHandler.java
@@ -1,36 +1,0 @@
-package com.mossy.boundedContext.global.security;
-
-import com.mossy.global.rsData.RsData;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import tools.jackson.databind.ObjectMapper;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class RestAccessDeniedHandler implements AccessDeniedHandler {
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void handle(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            AccessDeniedException accessDeniedException
-    ) throws IOException {
-
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-
-        objectMapper.writeValue(
-                response.getWriter(),
-                RsData.fail("F-403", "접근 권한이 없습니다.")
-        );
-    }
-}
