--- member/src/main/java/com/mossy/boundedContext/global/security/UserDetailsServiceImpl.java
@@ -1,37 +1,0 @@
-package com.mossy.boundedContext.global.security;
-
-import com.mossy.boundedContext.domain.user.User;
-import com.mossy.boundedContext.exception.DomainException;
-import com.mossy.boundedContext.exception.ErrorCode;
-import com.mossy.member.out.user.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public UserDetails loadUserByUsername(String email) {
-        User user = userRepository.findByEmailWithRoles(email)
-                .orElseThrow(()  -> new DomainException(ErrorCode.USER_NOT_FOUND));
-
-        return new UserDetailsImpl(user);
-    }
-
-    @Cacheable(value = "USER_DETAILS", key = "#userId")
-    @Transactional(readOnly = true)
-    public UserDetailsImpl loadUserById(Long userId) {
-        User user = userRepository.findByIdWithRoles(userId)
-                .orElseThrow(()  -> new DomainException(ErrorCode.USER_NOT_FOUND));
-
-        return new UserDetailsImpl(user);
-    }
-}
