--- member/src/main/java/com/mossy/boundedContext/in/dto/response/SellerApproveResponse.java
@@ -1,8 +1,0 @@
-package com.mossy.shared.member.dto.request;
-
-public record SellerApproveResponse(
-        Long sellerId,
-        String accessToken,
-        String refreshToken
-) {
-}
