+++ member/src/main/java/com/mossy/member/exception/SuccessCode.java
@@ -1,0 +1,7 @@
+package com.mossy.member.exception;
+
+import lombok.Getter;
+
+@Getter
+public enum SuccessCode {
+}
