+++ member/src/main/java/com/mossy/member/in/dto/request/SellerRequestCreateRequest.java
@@ -1,0 +1,22 @@
+package com.mossy.member.in.dto.request;
+
+import com.mossy.shared.member.domain.enums.SellerType;
+import jakarta.persistence.Column;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
+
+public record SellerRequestCreateRequest(
+        @NotNull SellerType sellerType,
+        @NotBlank String storeName,
+        @NotBlank String businessNum,
+        @NotBlank String representativeName,
+        String contactEmail,
+        String contactPhone,
+        @NotBlank String address1,
+        String address2,
+        @NotNull BigDecimal latitude,
+        @NotNull BigDecimal longitude
+) {
+}
