+++ member/src/main/java/com/mossy/member/in/dto/response/SellerApproveResponse.java
@@ -1,0 +1,8 @@
+package com.mossy.member.in.dto.response;
+
+public record SellerApproveResponse(
+        Long sellerId,
+        String accessToken,
+        String refreshToken
+) {
+}
